+++ templates/extensions/usd_viewer.messaging/template/{{python_module_path}}/tests/__init__.py
@@ -8,4 +8,4 @@
 # without an express license agreement from NVIDIA CORPORATION or
 # its affiliates is strictly prohibited.
 
-# from .messaging_tests import *
+from .messaging_tests import *