+++ templates/extensions/usd_viewer.setup/template/{{python_module_path}}/tests/test_app_extensions.py
@@ -13,7 +13,9 @@ from omni.kit.test import AsyncTestCase
 
 
 class TestUSDViewerExtensions(AsyncTestCase):
-    # NOTE: Function pulled to remove dependency from omni.kit.core.tests
+    """Class to hold USD Viewer setup extension tests."""
+    # NOTE: Function pulled to remove dependency from omni.kit.core.tests"""
+
     def _validate_extensions_load(self):
         failures = []
         manager = omni.kit.app.get_app().get_extension_manager()